% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_safe_set_names.R
\name{bq_safe_set_names}
\alias{bq_safe_set_names}
\title{Set BigQuery Safe Names}
\usage{
bq_safe_set_names(names)
}
\arguments{
\item{names}{list of dataframe column names}
}
\value{
list of BQ safe column name
}
\description{
Remove bad characters so colnames are safe for loading into BigQuery
Original Author: Mark Edmondson https://github.com/MarkEdmondson1234
}
