% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/extract_ga_data.R
\name{extract_ga_data}
\alias{extract_ga_data}
\title{Extract GA Data}
\usage{
extract_ga_data(ga_ids, extract_dates = NULL, extract_metrics = NULL,
  extract_dimensions = NULL, dim_list_map = NULL,
  dim_filter_string = NULL, outfile_dim_colname = NULL,
  outfile_tz_col = FALSE, outfile_directory = NULL)

extract_ga_data(ga_ids, extract_dates = NULL, extract_metrics = NULL,
  extract_dimensions = NULL, dim_list_map = NULL,
  dim_filter_string = NULL, outfile_dim_colname = NULL,
  outfile_tz_col = FALSE, outfile_directory = NULL)
}
\arguments{
\item{outfile_directory}{}

\item{outfile_directory}{}
}
\value{
a dataframe object

a dataframe object
}
\description{
A function for looping trhough a list of
GA Views to extract custom dimensions implemented across views

A function for looping trhough a list of
GA Views to extract custom dimensions implemented across views
}
\concept{extract functions}
