% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ga_data.R
\name{load_ga_data}
\alias{load_ga_data}
\title{Load GA Data to BQ
for use in Shiny Apps to preserve against data loss
and optimize Docker/GCR workflow by not storing data locally}
\usage{
load_ga_data(ga_ids, outfile_directory = NULL, create, overwrite = NULL)
}
\arguments{
\item{overwrite}{}
}
\description{
Load GA Data to BQ
for use in Shiny Apps to preserve against data loss
and optimize Docker/GCR workflow by not storing data locally
}
