% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{safe_set_names}
\alias{safe_set_names}
\title{Safe set names}
\usage{
safe_set_names(data_frame, lookup_names)
}
\arguments{
\item{data_frame}{Data.frame to change the names of}

\item{lookup_names}{A named vector of the names of old_name = new_name}
}
\value{
A character vector of new names, or error if any NA's
}
\description{
Will error if there is a problem setting names, which is better than silently returning NA
Original Author: Mark Edmondson https://github.com/MarkEdmondson1234
}
