% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gce_vm.R
\name{setup_gce_vm}
\alias{setup_gce_vm}
\title{Setup GCE VM.
A helper function that sets up a GCE VM, waits for a period of time you set
then opens the shiny app in your browser}
\usage{
setup_gce_vm(gce_vm_name, gce_vm_temp, gce_vm_type, image_tag,
  app_directory, wait_seconds)
}
\arguments{
\item{gce_vm_name}{STRING}

\item{gce_vm_type}{STRING}

\item{image_tag}{STRING}

\item{app_directory}{STRING}

\item{wait_second}{INTEGER of seconds to wait before opening the app's url}
}
\value{
a dataframe "vm" wtih GCE instance info
}
\description{
Setup GCE VM.
A helper function that sets up a GCE VM, waits for a period of time you set
then opens the shiny app in your browser
}
